errors removed

+++ delete.php
@@ -1,9 +1,11 @@
 <?php
 require('db.php');
-$id = $_POST['id'];
-$query = 'DELETE FROM user WHERE id =' . $id;
-$data = mysqli_query($con, $query);
+if (isset($_POST['id'])) {
+    $id = $_POST['id'];
+    $query = 'DELETE FROM user WHERE id =' . $id;
+    $data = mysqli_query($con, $query);
 
-if ($data) {
-    echo "deleted";
+    if ($data) {
+        echo "deleted";
+    }
 }
